cukinia/kernel.conf: remove gcc test
Since GCC 13, CONFIG_CC_HAS_AUTO_VAR_INIT_ZERO is kernel hardening
config enabled, and is not compatible with
CONFIG_GCC_PLUGIN_STRUCTLEAK_BYREF_ALL, which is not enabled anymore in
Scarthgap. This test has no reason to be kept, so removing it and
replace it with CONFIG_CC_HAS_AUTO_VAR_INIT_ZERO test.

Change-Id: I4730c3875cef729f9d62111fe810a388c0d91bc7
Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>