cluster-glue: system-config-cluster: set USERADD_DEPENDS
Some builds (CI or manual) encounter an error were the
do_package_setscene task of recipes depending on groups
created by other recipes fail because useradd can't find the groups.
This error was already reported on the yocto bugzilla [1].

With scarthgap, commit [2] of poky introduced the new variable
USERADD_DEPENDS to fix this issue. It adds the listed recipes
to DEPENDS and creates additional task dependencies to fix the issue.

Use this variable with cluster-glue bbappend and system-config-cluster
recipe, as the problem was frequently encountered with these recipes.

[1]: https://bugzilla.yoctoproject.org/show_bug.cgi?id=13419
[2]: https://git.yoctoproject.org/poky/commit/?id=ecef665062be55fcfa0915216335d08883aa86f7

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>