system-config: add glibc-utils RDEPENDS to system-config-security
system-config-security is a package which is include when the SEAPATH
security feature is enabled. This feature requires the `getent` command
which is provided by the `libc6-utils` package. This commit adds the
`glibc-utils` package to the RDEPENDS of the `system-config-security`
package. The glic-utils package is the glibc implementation of the
`libc6-utils` package

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>