seapath-host-common.inc: install python3-setup-ovs and system-config-ovs
OpenVSwitch and python3-setup-ovs configuration was moved out from
system-config-cluster to its own package to configure it on
standalone image as well. SEAPATH Debian standalone does install
OpenVSwitch and python3-setup-ovs, so we want to also install them
on SEAPATH Yocto standalone for harmonization.

Explicitly install python3-setup-ovs and system-config-ovs in
seapath-host-common.inc so both cluster and standalone SEAPATH Yocto are
packages with python3-setup-ovs, OpenVSwitch and the SEAPATH
configuration.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>